pickup/requirements/ubuntu.txt
apturl==0.5.2
asn1crypto==0.22.0
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==3.1.0
blinker==1.3
Brlapi==0.6.5
certifi==2018.10.15
chardet==3.0.4
Click==7.0
colorama==0.4.1
command-not-found==0.3
cryptography==1.9
csvkit==0.9.1
cupshelpers==1.0
cycler==0.10.0
decorator==4.3.2
defer==1.0.6
defusedxml==0.5.0
distro-info==0.17
dj-database-url==0.5.0
Django==2.1.7
django-crispy-forms==1.7.2
django-friendship==1.8.0
django-geojson==2.12.0
django-heroku==0.3.1
docopt==0.6.2
entrypoints==0.3
Flask==0.12.2
html5lib==0.999999999
httplib2==0.9.2
idna==2.7
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==0.24
jdcal==1.0
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
keyring==10.4.0
keyrings.alt==2.2
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.5
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.0.0
lxml==4.0.0
Mako==1.0.7
Markdown==2.6.9
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
notebook==5.7.6
numpy==1.16.2
oauth==1.0.1
olefile==0.44
openpyxl==2.3.0
pandas==0.24.2
pandoc==1.0.2
pandocfilters==1.4.2
parso==0.3.4
path-and-address==2.0.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==4.1.1
ply==3.11
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psycopg2==2.7.7
ptyprocess==0.6.0
py==1.4.34
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
pygobject==3.24.1
pyinotify==0.9.6
pyOpenSSL==16.2.0
pyparsing==2.3.1
pyrsistent==0.14.11
pytest==3.1.3
python-apt==1.4.0b3
python-dateutil==2.8.0
python-debian==0.1.30
python-igraph==0.7.1.post6
pytz==2018.9
pyxdg==0.25
PyYAML==3.12
pyzmq==18.0.1
qtconsole==4.4.3
reportlab==3.4.0
requests==2.20.1
scikit-learn==0.20.3
scipy==1.2.1
screen-resolution-extra==0.0.0
seaborn==0.9.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.11.1
six==1.12.0
sklearn==0.0
SQLAlchemy==1.1.9
system-service==0.3
systemd-python==234
terminado==0.8.1
testpath==0.4.2
tornado==6.0.1
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2.dev0
whitenoise==4.1.2
widgetsnbextension==3.4.2
xkit==0.0.0
xlrd==1.0.0
zope.interface==4.3.2
